// frontend/src/app/core/api/generated/model/page-response-product-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductResponse } from './product-response';


export interface PageResponseProductResponse { 
    content?: Array<ProductResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    first?: boolean;
    last?: boolean;
}